2nd commit

+++ src/services/post.service.js
@@ -1,8 +1,9 @@
-import prisma from "../client";
-import { FileService } from "./file.service";
+import prisma from "../client.js";
+import ApiError from "../exceptions/api.error.js";
+import FileService from "./file.service.js";
 
-export class PostService {
-    async createPost(content, media, authorId) {
+class PostService {
+    async createPost(contnet, media, authorId) {
         return await prisma.post
             .create({
                 data: {
@@ -14,8 +15,11 @@ export class PostService {
             .then((data) => data.id);
     }
 
-    async getAllPost(page) {
+    async getAllPost(page, authorId) {
         return await prisma.post.findMany({
+            where: {
+                authorId,
+            },
             skip: 20 * (page - 1),
             take: 20,
         });
@@ -29,25 +33,36 @@ export class PostService {
         });
     }
 
-    async updatePost(id, content, media) {
-        return await prisma.post.update({
-            where: {
-                id,
-            },
-            data: {
-                contnet,
-                media,
-            },
-        });
+    async updatePost(id, contnet, media, authorId) {
+        const post = await this.getPostById(id);
+        if (post.authorId === authorId) {
+            return await prisma.post.update({
+                where: {
+                    id,
+                },
+                data: {
+                    contnet,
+                    media,
+                },
+            });
+        } else {
+            throw ApiError.BadRequest("Вы не можете редачить этот пост");
+        }
     }
 
-    async deletePost(id) {
+    async deletePost(id, authorId) {
         const post = await this.getPostById(id);
-        if (post.media.length) {
-            await Promise.all(post.media.map(async (item) => await FileService.deleteFile(item)));
+        if (post.authorId === authorId) {
+            if (post.media.length) {
+                await Promise.all(post.media.map(async (item) => await FileService.deleteFile(item)));
+            }
+            return await prisma.post.delete({
+                where: { id },
+            });
+        } else {
+            throw ApiError.BadRequest("Вы не можете удалить этот пост");
         }
-        return await prisma.post.delete({
-            where: { id },
-        });
     }
 }
+
+export default new PostService();
